refactor(query): extract tool execution helpers into runner/tools.rs
